added children calc prop to individual
