+++ src/encoded/types/__init__.py
@@ -30,72 +30,6 @@ def includeme(config):
 
 
 @collection(
-    name='individuals',
-    unique_key='accession',
-    properties={
-        'title': 'Individuals',
-        'description': 'Listing of Individuals',
-    })
-class Individual(Item):
-    item_type = 'individual'
-    name_key = 'accession'
-    schema = load_schema('encoded:schemas/individual.json')
-
-    embedded_list = [
-        'father.is_deceased',
-        'father.sex',
-        'mother.is_deceased',
-        'mother.sex'
-    ]
-
-    # rev = {
-    #     'cohorts_proband': ('Cohort', 'proband'),
-    #     'cohorts_affiliate' : ('Cohort', 'trio.individual')
-    # }
-
-    # @calculated_property(schema={
-    #     "title": "Cohorts (proband)",
-    #     "description": "Cohorts that this individual is a proband of",
-    #     "type": "array",
-    #     "items": {
-    #         "title": "Cohort",
-    #         "type": "string",
-    #         "linkTo": "Cohort"
-    #     }
-    # })
-    # def cohorts_proband(self, request):
-    #     return self.rev_link_atids(request, "cohorts_proband")
-
-    # @calculated_property(schema={
-    #     "title": "Cohorts (affiliated)",
-    #     "description": "Cohorts that this individual is affiliated with",
-    #     "type": "array",
-    #     "items": {
-    #         "title": "Cohort",
-    #         "type": "string",
-    #         "linkTo": "Cohort"
-    #     }
-    # })
-    # def cohorts_affiliate(self, request):
-    #     return self.rev_link_atids(request, "cohorts_affiliate")
-
-    @calculated_property(schema={
-        "title": "Display Title",
-        "description": "A calculated title for every object in 4DN",
-        "type": "string"
-    })
-    def display_title(self, accession, bgm_id=None, other_id=None):
-        """Use bgm_id, other_id, or accession (in that order)"""
-        if bgm_id:
-            title = bgm_id
-        elif other_id:
-            title = '%s (%s)' % (other_id['id'], other_id['id_source'])
-        else:
-            title = accession
-        return title
-
-
-@collection(
     name='samples',
     unique_key='accession',
     properties={
